Se arreglo funcion db y id
