+++ api/src/models/Videogame.js
@@ -40,17 +40,6 @@ module.exports = (sequelize) => {
             type: DataTypes.BOOLEAN,
             allowNull: false,
         },
-        // genre: {
-        //     type: DataTypes.ARRAY(DataTypes.STRING),
-        //     allowNull: false,   
-        //     defaultValue: ["Has no genres"]
-                     
-        //   },
-        // tag: {
-        //     type: DataTypes.ARRAY(DataTypes.STRING),
-        //     allowNull: false,
-        //     defaultValue: ["Has no tags"]
-        // },
         short_screenshots: {
             type: DataTypes.ARRAY(DataTypes.TEXT),
             allowNull: false,
